Created new Directory Java_Programming_Week7 containing files such as:
i. LogEntry.java: class to specify the blueprint of a log entry of a web server
ii. LogAnalyzer.java: Reads a log file from system and creates an ArrayList of LogEntry for entries in file
iii. WebLogParser.java: Helps in parsing a single entry of log file into a LogEntry object
iv. Tester.java: to test the LogAnalyzer and LogEntry class
v. short-test_log which contains few log entries of a web server